+++ src/phold/__init__.py
@@ -193,6 +193,25 @@ def compare_options(func):
     return func
 
 
+"""
+compare only options used for genbank/genome FASTA input (i.e. not proteins-compare)
+"""
+
+
+def compare_no_proteins_options(func):
+    """compare command line args"""
+    options = [
+        click.option(
+            "--clinker",
+            is_flag=True,
+            help="Create output to use consistent PHROG categories and Phold plot colours for clinker.\nCreates gene_functions.csv for use with -gf and colour_map.csv for use with -cm clinker options.",
+        )
+    ]
+    for option in reversed(options):
+        func = option(func)
+    return func
+
+
 @click.group()
 @click.help_option("--help", "-h")
 @click.version_option(get_version(), "--version", "-V")
@@ -219,6 +238,7 @@ run command
 @common_options
 @predict_options
 @compare_options
+@compare_no_proteins_options
 def run(
     ctx,
     input,
@@ -242,6 +262,7 @@ def run(
     save_per_protein_embeddings,
     only_representatives,
     ultra_sensitive,
+    clinker,
     **kwargs,
 ):
     """phold predict then comapare all in one - GPU recommended"""
@@ -274,6 +295,7 @@ def run(
         "--save_per_protein_embeddings": save_per_protein_embeddings,
         "--only_representatives": only_representatives,
         "--ultra_sensitive": ultra_sensitive,
+        "--clinker": clinker,
     }
 
     # initial logging etc
@@ -323,7 +345,7 @@ def run(
         prefix,
         predictions_dir=output,
         structures=False,
-        structure_dir=output,
+        structure_dir=None,
         logdir=logdir,
         filter_structures=False,
         remote_flag=True,
@@ -333,6 +355,7 @@ def run(
         max_seqs=max_seqs,
         only_representatives=only_representatives,
         ultra_sensitive=ultra_sensitive,
+        clinker=clinker,
     )
 
     # cleanup the temp files
@@ -476,6 +499,7 @@ compare command
 )
 @common_options
 @compare_options
+@compare_no_proteins_options
 def compare(
     ctx,
     input,
@@ -496,6 +520,7 @@ def compare(
     max_seqs,
     only_representatives,
     ultra_sensitive,
+    clinker,
     **kwargs,
 ):
     """Runs Foldseek vs phold db"""
@@ -526,6 +551,7 @@ def compare(
         "--max_seqs": max_seqs,
         "--only_representatives": only_representatives,
         "--ultra_sensitive": ultra_sensitive,
+        "--clinker": clinker,
     }
 
     # initial logging etc
@@ -561,6 +587,7 @@ def compare(
         max_seqs=max_seqs,
         only_representatives=only_representatives,
         ultra_sensitive=ultra_sensitive,
+        clinker=clinker,
     )
 
     # cleanup the temp files
@@ -632,7 +659,7 @@ def proteins_predict(
     }
 
     # initial logging etc
-    start_time = begin_phold(params, "protein-predict")
+    start_time = begin_phold(params, "proteins-predict")
 
     # check the database is installed
     database = validate_db(database, DB_DIR)
@@ -811,7 +838,7 @@ def proteins_compare(
             logger.error(
                 f"Your input file {input} is likely not a amino acid FASTA file. Please check this."
             )
-        for record in SeqIO.parse(handle, "fasta"):
+        for record in records:
             prot_id = record.id
             feature_location = FeatureLocation(0, len(record.seq))
             # Seq needs to be saved as the first element in list hence the closed brackets [str(record.seq)]
@@ -851,6 +878,7 @@ def proteins_compare(
         max_seqs=max_seqs,
         only_representatives=only_representatives,
         ultra_sensitive=ultra_sensitive,
+        clinker=False,  # never create clinker output for proteins-compare as it doesnt make sense
     )
 
     # cleanup the temp files
@@ -879,6 +907,7 @@ remote command
 )
 @common_options
 @compare_options
+@compare_no_proteins_options
 def remote(
     ctx,
     input,
@@ -895,6 +924,7 @@ def remote(
     max_seqs,
     only_representatives,
     ultra_sensitive,
+    clinker,
     **kwargs,
 ):
     """Uses Foldseek API to run ProstT5 then Foldseek locally"""
@@ -920,6 +950,7 @@ def remote(
         "--max_seqs": max_seqs,
         "--only_representatives": only_representatives,
         "--ultra_sensitive": ultra_sensitive,
+        "--clinker": clinker,
     }
 
     # initial logging etc
@@ -988,7 +1019,7 @@ def remote(
         prefix,
         predictions_dir=output,
         structures=False,
-        structure_dir=output,
+        structure_dir=None,
         logdir=logdir,
         filter_structures=False,
         remote_flag=True,
@@ -998,6 +1029,7 @@ def remote(
         max_seqs=max_seqs,
         only_representatives=only_representatives,
         ultra_sensitive=ultra_sensitive,
+        clinker=clinker,
     )
 
     # cleanup the temp files