+++ src/phold/subcommands/compare.py
@@ -40,6 +40,7 @@ def subcommand_compare(
     max_seqs: int,
     only_representatives: bool,
     ultra_sensitive: bool,
+    clinker: bool,
 ) -> bool:
     """
     Compare 3Di or PDB structures to the Phold DB
@@ -65,6 +66,7 @@ def subcommand_compare(
         max_seqs (int): Maximum results per query sequence allowed to pass the prefilter for foldseek.
         only_representatives (bool): Whether to search against representatives only (turn off --cluster-search 1)
         ultra_sensitive (bool): Whether to skip foldseek prefilter for maximum sensitivity
+        clinker (bool): If True, then outputs gene_functions.csv for use with -gf and colour_map.csv for use with -cm clinker options
 
     Returns:
         bool: True if sub-databases are created successfully, False otherwise.
@@ -72,7 +74,7 @@ def subcommand_compare(
 
     if predictions_dir is None and structures is False:
         logger.error(
-            f"You did not specify --structure_dir or --structures. Please check "
+            f"You did not specify --predictions_dir or --structures. Please check "
         )
 
     if structures and structure_dir is None:
@@ -207,7 +209,7 @@ def subcommand_compare(
                 logger.error(
                     f"The AA CDS file {fasta_aa_input} does not exist. Please run phold predict and/or check the prediction directory {predictions_dir}"
                 )
-    ## write the AAs to file if structures is true
+    ## write the AAs to file if structures is true because can't just copy from prediction_dir
     else:
         ## write the CDS to file
         logger.info(f"Writing the AAs to file {fasta_aa}.")
@@ -217,8 +219,15 @@ def subcommand_compare(
 
                 # writes the CDS to file
                 for seq_id, cds_feature in aa_contig_dict.items():
-                    out_f.write(f">{record_id}:{seq_id}\n")
-                    out_f.write(f"{cds_feature.qualifiers['translation'][0]}\n")
+                    # if proteins, don't want the 'proteins:' as CDS id
+                    if proteins_flag:
+                        header = f">{seq_id}\n"
+                        seq = f"{cds_feature.qualifiers['translation']}\n"
+                    else:  # if genbank entry need to take the first seq as it is parsed as a list
+                        header = f">{record_id}:{seq_id}\n"
+                        seq = f"{cds_feature.qualifiers['translation'][0]}\n"
+                    out_f.write(header)
+                    out_f.write(seq)
 
     ############
     # create foldseek db
@@ -245,6 +254,7 @@ def subcommand_compare(
             logdir,
             prefix,
             filter_structures,
+            proteins_flag,
         )
     else:
         generate_foldseek_db_from_aa_3di(
@@ -410,9 +420,43 @@ def subcommand_compare(
 
     # save
     merged_df_path: Path = Path(output) / f"{prefix}_per_cds_predictions.tsv"
-
     merged_df.to_csv(merged_df_path, index=False, sep="\t")
 
+    # clinker output
+
+    if clinker:
+        clinker_dir = Path(output) / "clinker"
+        clinker_dir.mkdir(parents=True, exist_ok=True)
+
+        # gf
+        clinker_gf_df = merged_df[["cds_id", "function"]].copy()
+        clinker_gf_path: Path = Path(clinker_dir) / f"{prefix}_gene_functions.csv"
+        clinker_gf_df.to_csv(clinker_gf_path, index=False, sep=",", header=False)
+
+        # cm
+        clinker_cm_df = merged_df[["function"]].copy()
+        clinker_cm_path: Path = Path(clinker_dir) / f"{prefix}_colourmap.csv"
+
+        # from phold plot
+        function_to_colour_dict = {
+            "unknown function": "#AAAAAA",
+            "other": "#4deeea",
+            "tail": "#74ee15",
+            "transcription regulation": "#ffe700",
+            "DNA, RNA and nucleotide metabolism": "#f000ff",
+            "lysis": "#001eff",
+            "moron, auxiliary metabolic gene and host takeover": "#8900ff",
+            "integration and excision": "#E0B0FF",
+            "head and packaging": "#ff008d",
+            "connector": "#5A5A5A",
+        }
+
+        clinker_cm_df["colour"] = clinker_cm_df["function"].map(function_to_colour_dict)
+        # remove dupes
+        clinker_cm_df = clinker_cm_df.drop_duplicates()
+        clinker_cm_df.to_csv(clinker_cm_path, index=False, sep=",", header=False)
+
+    # sub dbs output
     # save vfdb card acr defensefinder hits with more metadata
     sub_dbs_created = create_sub_db_outputs(merged_df, database, output)